Docker tag must be lowercase; hardcoding for my repo; fine for main project repos
Need to hardcode for my repos since my docker username is lowercase but github username is uppercase.
Since I want to use input.repo for both repo name checks and docker image tags, it's better to not change code.
It's just that it's not working for my repos.
Since e-mission project org owner and repo names are lowercase, and repo names as well lowercase, it should be fine.